refactor(websocket) :  websocket 파일 수정
